# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  productId: string;
  name: string;
  quantity: number;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
}